fix(doctor): preserve cross-platform startup checks
